addressing PR feedback'
