+++ sdk/core/Azure.Core/tests/TestFramework/RecordTransport.cs
@@ -73,12 +73,10 @@ namespace Azure.Core.Testing
                 RequestMethod = request.Method,
                 Request =
                 {
-                    Headers = new SortedDictionary<string, string[]>(StringComparer.OrdinalIgnoreCase),
                     Body = ReadToEnd(request.Content),
                 },
                 Response =
                 {
-                    Headers = new SortedDictionary<string, string[]>(StringComparer.OrdinalIgnoreCase),
                     Body = ReadToEnd(response),
                 },
                 StatusCode = response.Status
